+++ src/Defaults.ts
@@ -138,18 +138,11 @@ const columnStyleDefaults = {
 };
 
 const dateFormats = [
-  {
-    text: 'YYYY-MM-DD HH:mm:ss',
-    value: 'YYYY-MM-DD HH:mm:ss',
-  },
-  {
-    text: 'MM/DD/YY h:mm:ss a',
-    value: 'MM/DD/YY h:mm:ss a',
-  },
-  {
-    text: 'MMMM D, YYYY LT',
-    value: 'MMMM D, YYYY LT',
-  },
+  { text: 'YYYY-MM-DD HH:mm:ss', value: 'YYYY-MM-DD HH:mm:ss' },
+  { text: 'YYYY-MM-DD HH:mm:ss.SSS', value: 'YYYY-MM-DD HH:mm:ss.SSS' },
+  { text: 'MM/DD/YY h:mm:ss a', value: 'MM/DD/YY h:mm:ss a' },
+  { text: 'MMMM D, YYYY LT', value: 'MMMM D, YYYY LT' },
+  { text: 'YYYY-MM-DD', value: 'YYYY-MM-DD' },
 ];
 
 const columnTypes = [
