--- src/Formatter.test.ts
@@ -1,5 +1,0 @@
-describe('Formatter test', () => {
-  it.skip('plugin should be defined', () => {
-    expect(true).toBeTruthy();
-  });
-});
